# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Orbo Video Alarm System
    description: Open source video alarm system with motion detection and Telegram integration
    version: "1.0"
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /api/v1/cameras/{id}/activate:
        post:
            tags:
                - camera
            summary: activate camera
            description: Activate camera for motion detection
            operationId: camera#activate
            parameters:
                - name: id
                  in: path
                  description: Camera ID
                  required: true
                  type: string
                  format: uuid
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CameraActivateResponseBody'
                        required:
                            - id
                            - name
                            - device
                            - status
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CameraActivateNotFoundResponseBody'
                        required:
                            - message
                            - id
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/CameraActivateInternalResponseBody'
                        required:
                            - message
            schemes:
                - http
    /api/v1/cameras/{id}/alerts/disable:
        post:
            tags:
                - camera
            summary: disable_alerts camera
            description: Disable alerts for this camera. Detection pipeline still runs for bounding boxes but no events are created or notifications sent.
            operationId: camera#disable_alerts
            parameters:
                - name: id
                  in: path
                  description: Camera ID
                  required: true
                  type: string
                  format: uuid
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CameraDisableAlertsResponseBody'
                        required:
                            - id
                            - name
                            - device
                            - status
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CameraDisableAlertsNotFoundResponseBody'
                        required:
                            - message
                            - id
            schemes:
                - http
    /api/v1/cameras/{id}/alerts/enable:
        post:
            tags:
                - camera
            summary: enable_alerts camera
            description: Enable alerts for this camera. Detection pipeline will create events and send notifications.
            operationId: camera#enable_alerts
            parameters:
                - name: id
                  in: path
                  description: Camera ID
                  required: true
                  type: string
                  format: uuid
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CameraEnableAlertsResponseBody'
                        required:
                            - id
                            - name
                            - device
                            - status
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CameraEnableAlertsNotFoundResponseBody'
                        required:
                            - message
                            - id
            schemes:
                - http
    /api/v1/cameras/{id}/deactivate:
        post:
            tags:
                - camera
            summary: deactivate camera
            description: Deactivate camera
            operationId: camera#deactivate
            parameters:
                - name: id
                  in: path
                  description: Camera ID
                  required: true
                  type: string
                  format: uuid
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CameraDeactivateResponseBody'
                        required:
                            - id
                            - name
                            - device
                            - status
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CameraDeactivateNotFoundResponseBody'
                        required:
                            - message
                            - id
            schemes:
                - http
    /api/v1/cameras/{id}/frame:
        get:
            tags:
                - camera
            summary: capture camera
            description: Capture a single frame from camera as base64
            operationId: camera#capture
            parameters:
                - name: id
                  in: path
                  description: Camera ID
                  required: true
                  type: string
                  format: uuid
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CameraCaptureResponseBody'
                        required:
                            - data
                            - content_type
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CameraCaptureNotFoundResponseBody'
                        required:
                            - message
                            - id
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/CameraCaptureInternalResponseBody'
                        required:
                            - message
            schemes:
                - http
...
definitions:
...
    CameraActivateInternalResponseBody:
        title: CameraActivateInternalResponseBody
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: Quo qui.
        description: Failed to activate camera
        example:
            message: Ut soluta.
        required:
            - message
    CameraActivateNotFoundResponseBody:
        title: CameraActivateNotFoundResponseBody
        type: object
        properties:
            id:
                type: string
                description: Resource ID
                example: Nesciunt consequuntur autem adipisci.
            message:
                type: string
                description: Error message
                example: Aut corporis numquam sit expedita.
        description: Camera not found
        example:
            id: Nulla illum exercitationem iusto temporibus quasi fugiat.
            message: Molestias et eius magni est sunt dolores.
        required:
            - message
            - id
    CameraActivateResponseBody:
        title: CameraActivateResponseBody
        type: object
        properties:
            alerts_enabled:
                type: boolean
                description: When false, detection pipeline still runs for bounding boxes but no events are created or alerts sent.
                example: false
            created_at:
                type: string
                description: Creation timestamp
                example: "2008-10-09T11:05:03Z"
                format: date-time
            device:
                type: string
                description: Camera device path (e.g., /dev/video0)
                example: Harum et quis non porro et.
            fps:
                type: integer
                description: Frames per second
                example: 3642552064099455358
                format: int64
            id:
                type: string
                description: Camera unique identifier
                example: ecddffac-f010-11f0-998b-5847ca7b8ce1
                format: uuid
            name:
                type: string
                description: Camera name
                example: Minus temporibus.
            resolution:
                type: string
                description: Camera resolution
                example: Ullam unde laboriosam quia ut.
            status:
                type: string
                description: Camera status
                example: error
                enum:
                    - active
                    - inactive
                    - error
        example:
            alerts_enabled: false
            created_at: "2011-01-09T01:40:20Z"
            device: Minus ullam commodi cumque.
            fps: 1265242636273343561
            id: ecde015b-f010-11f0-998b-5847ca7b8ce1
            name: Et rem ab.
            resolution: Magni quia.
            status: active
        required:
            - id
            - name
            - device
            - status
...
    CameraDisableAlertsNotFoundResponseBody:
        title: CameraDisableAlertsNotFoundResponseBody
        type: object
        properties:
            id:
                type: string
                description: Resource ID
                example: Sint quia ullam placeat rerum odio.
            message:
                type: string
                description: Error message
                example: Accusantium fugiat iusto provident exercitationem.
        description: Camera not found
        example:
            id: Quidem qui deleniti.
            message: Molestias quis iste natus.
        required:
            - message
            - id
    CameraDisableAlertsResponseBody:
        title: CameraDisableAlertsResponseBody
        type: object
        properties:
            alerts_enabled:
                type: boolean
                description: When false, detection pipeline still runs for bounding boxes but no events are created or alerts sent.
                example: false
            created_at:
                type: string
                description: Creation timestamp
                example: "1976-04-16T05:08:14Z"
                format: date-time
            device:
                type: string
                description: Camera device path (e.g., /dev/video0)
                example: Et veritatis possimus et.
            fps:
                type: integer
                description: Frames per second
                example: 8340487245942408587
                format: int64
            id:
                type: string
                description: Camera unique identifier
                example: ecde1472-f010-11f0-998b-5847ca7b8ce1
                format: uuid
            name:
                type: string
                description: Camera name
                example: Earum magnam earum quis vero nam neque.
            resolution:
                type: string
                description: Camera resolution
                example: Debitis laudantium eos.
            status:
                type: string
                description: Camera status
                example: error
                enum:
                    - active
                    - inactive
                    - error
        example:
            alerts_enabled: false
            created_at: "1998-11-05T04:33:38Z"
            device: Eligendi consequatur doloribus.
            fps: 4352859274004166387
            id: ecde184d-f010-11f0-998b-5847ca7b8ce1
            name: Voluptatem quo soluta quas.
            resolution: Perferendis pariatur tempora nulla aspernatur.
            status: error
        required:
            - id
            - name
            - device
            - status
    CameraEnableAlertsNotFoundResponseBody:
        title: CameraEnableAlertsNotFoundResponseBody
        type: object
        properties:
            id:
                type: string
                description: Resource ID
                example: Perspiciatis omnis sed aut.
            message:
                type: string
                description: Error message
                example: Dolorem eum.
        description: Camera not found
        example:
            id: Ullam pariatur.
            message: Sapiente est autem doloribus.
        required:
            - message
            - id
    CameraEnableAlertsResponseBody:
        title: CameraEnableAlertsResponseBody
        type: object
        properties:
            alerts_enabled:
                type: boolean
                description: When false, detection pipeline still runs for bounding boxes but no events are created or alerts sent.
                example: false
            created_at:
                type: string
                description: Creation timestamp
                example: "1998-08-29T02:34:17Z"
                format: date-time
            device:
                type: string
                description: Camera device path (e.g., /dev/video0)
                example: Temporibus inventore a corrupti quia voluptatibus.
            fps:
                type: integer
                description: Frames per second
                example: 1122283068990033033
                format: int64
            id:
                type: string
                description: Camera unique identifier
                example: ecde0be3-f010-11f0-998b-5847ca7b8ce1
                format: uuid
            name:
                type: string
                description: Camera name
                example: Voluptate architecto officiis voluptatem.
            resolution:
                type: string
                description: Camera resolution
                example: Porro maiores.
            status:
                type: string
                description: Camera status
                example: active
                enum:
                    - active
                    - inactive
                    - error
        example:
            alerts_enabled: true
            created_at: "2013-02-12T22:33:38Z"
            device: Sunt voluptatem.
            fps: 4914430661198993374
            id: ecde0fda-f010-11f0-998b-5847ca7b8ce1
            name: Natus consequatur quibusdam nisi.
            resolution: Delectus at dolorem illo.
            status: error
        required:
            - id
            - name
            - device
            - status
...